requirements.txt
streamlit
yfinance
seaborn
pandas
numpy
numerize

python-dotenv
google-cloud-storage